refactor: Common/Entities cleanup
