+++ JobScraper.Web/Features/CustomScrapers/Logic/TestFullScenario.cs
@@ -1,6 +1,7 @@
 using System.Text.Json;
 using ErrorOr;
 using JobScraper.Web.Common.Entities;
+using JobScraper.Web.Common.Models;
 using JobScraper.Web.Features.CustomScrapers.Models;
 using JobScraper.Web.Features.JobOffers.Scrape.Logic.Common;
 using JobScraper.Web.Modules.Persistence;
