+++ JobScraper.Web/Features/JobOffers/AddUpdateOffer.razor
@@ -3,6 +3,7 @@
 @using System.Web
 @using Blazored.FluentValidation
 @using JobScraper.Web.Common.Entities
+@using JobScraper.Web.Common.Models
 @using JobScraper.Web.Modules.Persistence
 @using Microsoft.EntityFrameworkCore
