+++ JobScraper.Web/Features/JobOffers/Applications.razor
@@ -2,6 +2,7 @@
 @using JobScraper.Web.Blazor.Components.MySelect
 @using JobScraper.Web.Common.Entities
 @using JobScraper.Web.Common.Extensions
+@using JobScraper.Web.Common.Models
 @using JobScraper.Web.Modules.Extensions
 @using JobScraper.Web.Modules.Persistence
 @using Microsoft.AspNetCore.Components.QuickGrid
