+++ JobScraper.Web/Features/JobOffers/JobOffers.razor
@@ -3,6 +3,7 @@
 @using JobScraper.Web.Blazor.Components.MySelect
 @using JobScraper.Web.Common.Entities
 @using JobScraper.Web.Common.Extensions
+@using JobScraper.Web.Common.Models
 @using JobScraper.Web.Features.JobOffers.Models
 @using JobScraper.Web.Modules.Extensions
 @using JobScraper.Web.Modules.Persistence
