+++ JobScraper.Web/Features/JobOffers/Scrape/Logic/Custom/CustomListScraper.cs
@@ -1,5 +1,6 @@
 using System.Text.Json;
 using JobScraper.Web.Common.Entities;
+using JobScraper.Web.Common.Models;
 using JobScraper.Web.Features.CustomScrapers.Models;
 using JobScraper.Web.Features.JobOffers.Scrape.Logic.Common;
 using JobScraper.Web.Modules.Persistence;
