+++ JobScraper.Web/Features/JobOffers/Scrape/Logic/Indeed/IndeedDetailsScraper.cs
@@ -1,4 +1,5 @@
 ﻿using JobScraper.Web.Common.Entities;
+using JobScraper.Web.Common.Models;
 using JobScraper.Web.Features.JobOffers.Scrape.Logic.Common;
 using JobScraper.Web.Modules.Persistence;
 using Microsoft.Extensions.Options;
