+++ JobScraper.Web/Features/JobOffers/Scrape/ScrapePage.razor
@@ -2,6 +2,7 @@
 @using BlazorBootstrap
 @using Blazored.FluentValidation
 @using JobScraper.Web.Common.Entities
+@using JobScraper.Web.Common.Models
 @using Microsoft.EntityFrameworkCore
 
 <NavigationLock ConfirmExternalNavigation="@PreventNavigation"
