+++ JobScraper.Web/Features/JobOffers/Scrape/ScrapePage.razor.cs
@@ -1,6 +1,7 @@
 ﻿using BlazorBootstrap;
 using Blazored.FluentValidation;
 using JobScraper.Web.Common.Entities;
+using JobScraper.Web.Common.Models;
 using JobScraper.Web.Features.AiSummary.Logic;
 using JobScraper.Web.Features.JobOffers.Scrape.Logic.Functions;
 using JobScraper.Web.Modules.Persistence;
